debugging

+++ Benchmarking/stats/tester/euler1/euler1.jl
@@ -4,7 +4,7 @@
 #Terminal: julia euler1.jl 1000
 limit = parse(Int, ARGS[1])
 function f1(n)
-  sum = 0
+  sum = UInt128(0)
   for i = 1:n-1
       if i%3 == 0 || i%5 == 0
         sum += i
